DEV-18456: Add updateUserGroupTranslations to the SDK
New UpdateUserGroupTranslationsDTO + CashbackClient::updateUserGroupTranslations
(PUT /users-groups/{id}/translations). StoreUserGroupDTO no longer carries the
translation maps.
