+++ src/App/DTO/Request/UserGroups/StoreUserGroupDTO.php
@@ -15,13 +15,7 @@ class StoreUserGroupDTO extends DTO
 
     protected string $name;
 
-    protected ?array $titleTranslations = null;
-
     protected ?int $transitionAmount = null;
-
-    protected ?array $benefitsDescriptionTranslations = null;
-
-    protected ?array $levelProgressTextTranslations = null;
 
     public function getAccountId(): int
     {
@@ -33,23 +27,8 @@ class StoreUserGroupDTO extends DTO
         return $this->name;
     }
 
-    public function getTitleTranslations(): ?array
-    {
-        return $this->titleTranslations;
-    }
-
     public function getTransitionAmount(): ?int
     {
         return $this->transitionAmount;
-    }
-
-    public function getBenefitsDescriptionTranslations(): ?array
-    {
-        return $this->benefitsDescriptionTranslations;
-    }
-
-    public function getLevelProgressTextTranslations(): ?array
-    {
-        return $this->levelProgressTextTranslations;
     }
 }
